+++ CHANGES.md
@@ -1,6 +1,7 @@
 0.2.0 (unreleased)
 =====
 - Add support for FLAC.
+- Use "bpm" instead of "tempo".
 
 0.1.0 (2023-02-08)
 =====