+++ CHANGES.md
@@ -1,7 +1,8 @@
 0.2.0 (unreleased)
 =====
 - Add support for FLAC.
-- Use "bpm" instead of "tempo".
+- id3v2: use "bpm" instead of "tempo".
+- id3v2: convert "tlen" to "duration".
 
 0.1.0 (2023-02-08)
 =====