+++ CHANGES.md
@@ -3,6 +3,7 @@
 - Add support for FLAC.
 - id3v2: use "bpm" instead of "tempo".
 - id3v2: convert "tlen" to "duration".
+- Fix charset conversion from utf16.
 
 0.1.0 (2023-02-08)
 =====