+++ CHANGES.md
@@ -4,6 +4,7 @@
 - id3v2: use "bpm" instead of "tempo".
 - id3v2: convert "tlen" to "duration".
 - Fix charset conversion from utf16.
+- Fixed MP4 parsing.
 
 0.1.0 (2023-02-08)
 =====