+++ CHANGES.md
@@ -3,7 +3,9 @@
 - Add support for FLAC.
 - id3v2: use "bpm" instead of "tempo".
 - id3v2: convert "tlen" to "duration".
+- id3v2: implement partial support for rendering.
 - Fix charset conversion from utf16.
+- Fixed MP4 parsing.
 
 0.1.0 (2023-02-08)
 =====